feat: enhance Kubernetes configuration and application metrics
- Updated deployment.yaml to include container specifications and readiness probes.
- Modified ingress.yaml to add SSL redirection annotation and routing rules.
- Enhanced kind-config.yaml with additional port mappings and kubeadm configuration.
- Updated service.yaml to define service ports and target ports.
- Expanded README.md with detailed project features, architecture, and setup instructions.
- Added python-dotenv to requirements.txt for environment variable management.
- Improved metrics.py to return Prometheus metrics with correct media type.
- Updated temperature.py to utilize a status map for temperature readings.
- Refactored integration tests in test_api.py to use FastAPI's TestClient.
- Adjusted unit tests in test_metrics.py to handle metric values as floats.

+++ tests/unit/test_metrics.py
@@ -45,7 +45,7 @@ def test_metrics_endpoint_increments_counter(client):
     response1 = client.get("/metrics")
     content1 = response1.text
     metrics_endpoint_requests1 = sum(
-        int(line.split(" ")[1])
+        float(line.split(" ")[1])
         for line in content1.split("\n")
         if line.startswith("hivebox_metrics_endpoint_requests_total")
     )
@@ -54,7 +54,7 @@ def test_metrics_endpoint_increments_counter(client):
     response2 = client.get("/metrics")
     content2 = response2.text
     metrics_endpoint_requests2 = sum(
-        int(line.split(" ")[1])
+        float(line.split(" ")[1])
         for line in content2.split("\n")
         if line.startswith("hivebox_metrics_endpoint_requests_total")
     )
